Request validations working

+++ app/Http/Controllers/OrdersController.php
@@ -8,6 +8,7 @@ use App\Models\Order;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Validator;
 
 class OrdersController extends Controller
 {
@@ -20,6 +21,14 @@ class OrdersController extends Controller
 
     public function create(Request $request): JsonResponse
     {
+        $validator = Validator::make($request->all(), [
+            'client_id' => 'exists:clients,id'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->messages());
+        }
+
         $order = new Order;
 
         $order->client_id = $request->client_id;
@@ -32,14 +41,14 @@ class OrdersController extends Controller
 
     public function show($id): JsonResponse
     {
-        $order = Order::find($id);
+        $order = Order::findOrFail($id);
 
         return response()->json($order);
     }
 
     public function update(Request $request, $id): JsonResponse
-    { 
-        $order = Order::find($id);
+    {
+        $order = Order::findOrFail($id);
         
         $order->note = $request->input('note');
                 
@@ -50,7 +59,7 @@ class OrdersController extends Controller
 
     public function destroy($id): JsonResponse
     {
-        $order = Order::find($id);
+        $order = Order::findOrFail($id);
         $order->delete();
 
         return response()->json('Pedido removido com sucesso');
@@ -58,19 +67,32 @@ class OrdersController extends Controller
 
     public function addProduct(Request $request, $id): JsonResponse
     {
-        $order = Order::find($id);
+        $validator = Validator::make($request->all(), [
+            'product_id' => 'exists:products,id'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->messages());
+        }
+
+        $order = Order::findOrFail($id);
         
         $productId = $request->input('product_id');
         $order->products()->attach($productId);
-
-        dd($order->products);
 
         return response()->json("Produto #{$productId} adicionado ao pedido #{$id}");
     }
 
     public function removeProduct(Request $request, $id): JsonResponse
     {
-        $order = Order::find($id);
+        $validator = Validator::make($request->all(), [
+            'product_id' => 'exists:products,id'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json($validator->messages());
+        }
+        $order = Order::findOrFail($id);
         
         $productId = $request->input('product_id');
         $order->products()->detach($productId);
